+++ packages/domain/src/storyboard/index.ts
@@ -4,3 +4,5 @@ export * from "./eligibility";
 export * from "./fingerprint";
 export * from "./ordering";
 export * from "./duration";
+export * from "./moderation";
+export * from "./prompt";